+++ js/pages/jsonData.js
@@ -6,7 +6,7 @@
  * be used for generic JSON web services. Often when creating custom pages
  * for DataFormsJS they will be based on this file.
  *
- * A class-based version of this file exsits at [js/pages/classes/JsonData.js]
+ * A class-based version of this file exists at [js/pages/classes/JsonData.js]
  * and can be used if you are developing an app with pages defined as classes
  * and need to extend from `JsonData`. If features are added to this file or
  * code is change the related code must also be changed in the class file.